Pedidos 80% falta detalle y filtros

+++ app/Http/Controllers/ProductoController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Entities\Catalogo;
 use App\Entities\Producto;
 use Illuminate\Http\Request;
 use App\Entities\GaleriaProducto;
+use DB;
 
 class ProductoController extends Controller
 {
@@ -68,12 +70,17 @@ class ProductoController extends Controller
         $producto->precio = $request['precio'];
         $producto->descripcion = $request['descripcion'];
         $producto->precio = $request['precio'];
+        $producto->total = $request['precio'];
         $producto->descuento = 0;
         $producto->iva = 19;
         $producto->catalogo = $request['catalogo'];
         $producto->marca = $request['marca'];
         
         if ($producto->save()) {
+            // Update cantidad catalogo.
+            $catalogo = Catalogo::find($request['catalogo']);
+            $catalogo->cantidad++;
+            $catalogo->save();
             // Save images.
             foreach ($request['imagenes'] as $key => $imagen) {
                 $name = "{$producto->referencia}-{$key}";
@@ -90,7 +97,6 @@ class ProductoController extends Controller
     }
 
     // Store image.
-    
     public function saveImage($image, $name, $id_catalogo){
         $extension = explode('/', explode(':', substr($image, 0, strpos($image, ';')))[1])[1];   // .jpg .png .pdf
         $replace = substr($image, 0, strpos($image, ',')+1);
@@ -109,28 +115,30 @@ class ProductoController extends Controller
         return $filename;
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Entities\Producto  $producto
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        $producto = Producto::find($id);
-        return $producto;
+
+    public function detalleProducto($id){
+        $producto = Producto::where('id_producto', $id)
+                    ->join('marcas', 'marca', '=', 'id_marca')
+                    ->first();
+        
+        $imagenes = GaleriaProducto::where('producto', $id)->get();
+        
+        foreach ($imagenes as $imagen) {
+            if($imagen->destacada){
+                $producto->destacada = url($imagen->image);
+            }
+            $imagen->image = url($imagen->image);
+        }
+
+        $producto->imagenes = $imagenes;
+        $response = [
+            'response' => 'success', 
+            'status' => 200,
+            'producto' => $producto
+        ];
+        return response()->json($response);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Entities\Producto  $producto
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Producto $producto)
-    {
-        //
-    }
 
     /**
      * Update the specified resource in storage.
@@ -139,10 +147,45 @@ class ProductoController extends Controller
      * @param  \App\Entities\Producto  $producto
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Producto $producto)
-    {
-        $producto->update($request->all());
-        return $producto;
+    public function update(Request $request, $id){
+        $request->validate([
+            'id_producto' => 'required',
+            'nombre' => 'required',
+            'codigo' => 'required',
+            'referencia' => 'required',
+            'stock' => 'required|numeric',
+            'marca' => 'required',
+            'descripcion' => 'required',
+            'precio' => 'required|numeric',
+            'catalogo' => 'required',
+        ]);
+        
+        $producto = Producto::find($request['id_producto']);
+        $producto->nombre = $request['nombre'];
+        $producto->codigo = $request['codigo'];
+        $producto->referencia = $request['referencia'];
+        $producto->stock = $request['stock'];
+        $producto->marca = $request['marca'];
+        $producto->precio = $request['precio'];
+        $producto->total = $request['precio'];
+        $producto->descripcion = $request['descripcion'];
+        
+        try {
+            $producto->save();
+            $response = [
+                'response' => 'success',
+                'status' => 200,
+                'message' => "Producto actualizado correctamente."
+            ];
+        } catch (\Exception $e) {
+            $response = [
+                'response' => 'error', 
+                'status' => 403,
+                'message' => $e->getMessage()
+            ];
+        }
+
+        return response()->json($response);
     }
 
     /**
@@ -151,9 +194,46 @@ class ProductoController extends Controller
      * @param  \App\Entities\Producto  $producto
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Producto $producto)
-    {
-        $producto->delete();
-        return $producto;
+    public function destroy($id){
+        
+        $validate_producto = DB::table('pedido_productos')->where('producto', $id)->exists();
+
+        if ($validate_producto) {
+            $response = [
+                'response' => 'error', 
+                'status' => 403,
+                'message' => "El producto tiene registro de pagos. No es posible que se elimine."
+            ];
+        }else{
+            $delete_images = GaleriaProducto::where('producto', $id)->delete();
+            if ($delete_images) {
+                $producto = Producto::find($id);
+
+                // Update cantidad catalogo.
+                $catalogo = Catalogo::find($producto->catalogo);
+                $catalogo->cantidad--;
+                if($catalogo->save()){
+                    $producto->delete();  
+                    $response = [
+                        'response' => 'success', 
+                        'status' => 200,
+                        'message' => "Producto eliminado."
+                    ];  
+                }else{
+                    $response = [
+                        'response' => 'error', 
+                        'status' => 403,
+                        'message' => "No se pudo eliminar el producto."
+                    ];  
+                }
+            }else{
+                $response = [
+                    'response' => 'error', 
+                    'status' => 403,
+                    'message' => "No se pudo eliminar el producto."
+                ];  
+            }
+        }
+        return response()->json($response);
     }
 }
